+++ src/pages/recipe-video/index.tsx
@@ -14,8 +14,8 @@ import {
   AlertCircle,
 } from 'lucide-react';
 import { Dialog, DialogContent, DialogTrigger } from '@/components/ui/dialog';
-import { fetchRecipeTeaching } from '@/services/api';
 import { RecipeTeachingResponse } from '@/types/api';
+import { fetchRecipeTeaching } from '@/services/recipes';
 
 /**
  * 食譜視頻頁面的 Props 介面
